Performance improvements Introduce Locality-Sensitive Hashing algorithm for faster searching of similar vectors and create a map for the lucene vector stores which speeds-up searching of vectors

+++ src/main/java/pitt/search/semanticvectors/collections/FileVectorStore.java
@@ -4,7 +4,7 @@ import com.ontotext.trree.io.ReadWriteFile;
 import pitt.search.semanticvectors.ObjectVector;
 import pitt.search.semanticvectors.vectors.Vector;
 
-import java.io.File;
+import java.io.*;
 import java.nio.ByteBuffer;
 import java.util.Enumeration;
 import java.util.HashMap;
@@ -21,14 +21,29 @@ public class FileVectorStore implements ModifiableVectorStore {
 	private Map<String, Long> keyToPos;
 	private Map<Long, String> posToKey;
 	private long size;
+	private boolean temp;
 
 	public FileVectorStore(File storageFile, VectorBufferConverter converter) {
+		this(storageFile, converter, true);
+	}
+
+	public FileVectorStore(File storageFile, VectorBufferConverter converter, boolean temp) {
 		this.bufferSizeInBytes = converter.getBufferSizeInBytes();
 		this.file = new ReadWriteFile(bufferSizeInBytes, storageFile, false);
 		this.converter = converter;
+		this.temp = temp;
 		keyToPos = new HashMap<>();
 		posToKey = new HashMap<>();
 		size = 0;
+
+		if (!temp && storageFile.exists() && storageFile.length() != 0) {
+			try {
+				restoreMaps();
+			} catch (IOException e) {
+				throw new RuntimeException(e);
+			}
+			size = keyToPos.size();
+		}
 	}
 
 	@Override
@@ -64,6 +79,7 @@ public class FileVectorStore implements ModifiableVectorStore {
 	public Enumeration<ObjectVector> getAllVectors() {
 		return new Enumeration<ObjectVector>() {
 			long currPos = 0;
+
 			@Override
 			public boolean hasMoreElements() {
 				return currPos < size;
@@ -93,10 +109,50 @@ public class FileVectorStore implements ModifiableVectorStore {
 	@Override
 	public void flush() {
 		file.flush();
+		if (!temp) {
+			persistKeyMap();
+		}
 	}
 
 	@Override
 	public void close() {
-		file.delete();
+		if (temp) {
+			file.delete();
+		} else {
+			persistKeyMap();
+		}
+	}
+
+	private void persistKeyMap() {
+		String storageFileName = file.getFileIdentifier().getAbsolutePath();
+		File keyMap = new File(storageFileName + ".keymap");
+		if (keyMap.exists())
+			keyMap.delete();
+
+		try (DataOutputStream dos = new DataOutputStream(new FileOutputStream(keyMap))) {
+			for (Map.Entry<String, Long> entry : keyToPos.entrySet()) {
+				dos.writeUTF(entry.getKey());
+				dos.writeLong(entry.getValue());
+			}
+		} catch (IOException e) {
+			throw new RuntimeException("Could not save keymap to file", e);
+		}
+	}
+
+	private void restoreMaps() throws IOException {
+		String storageFileName = file.getFileIdentifier().getAbsolutePath();
+		File keyMap = new File(storageFileName + ".keymap");
+		if (!keyMap.exists())
+			throw new RuntimeException("KeyMap file does not exist! " + keyMap.getAbsolutePath());
+
+		try (FileInputStream fis = new FileInputStream(keyMap);
+			 DataInputStream dis = new DataInputStream(fis)) {
+			while (fis.getChannel().position() < fis.getChannel().size()) {
+				String key = dis.readUTF();
+				long position = dis.readLong();
+				keyToPos.put(key, position);
+				posToKey.put(position, key);
+			}
+		}
 	}
 }
